backend/requirements.txt
flask
flask-cors
transformers
torch
Pillow
deep-translator
...

